Use :init for kele-mode
Using :config without :defer implied enabling kele-mode within an :init
block, but wasn't as clear in terms of intention. kele-mode needs to be
enabled in order to kele-dispatch to function, so it's now in an :init
block to be more clear. I've also omitted the positive integer argument,
as an ommitted arg from a non-interactive context also implies
enablement of the mode and looks cleaner in my opinion.