Change deprecated MacOS detection + eshell
Experimenting with smart shell display to automatically copy commands
that error in Eshell so they can be edited.

I also noticed the Emacs was telling me that the `IS-MAC` variable,
which comes from Doom, is deprecated as of V3. So I've followed the
deprecation notice and switched to the Emacs builtin: featurep.